Divide Navigation logic from MainActivity
