+++ app/src/main/java/com/example/androiddevchallenge/MainActivity.kt
@@ -22,13 +22,6 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavType
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.navArgument
-import androidx.navigation.compose.rememberNavController
-import com.example.androiddevchallenge.ui.screens.CatDetailScreen
-import com.example.androiddevchallenge.ui.screens.CatListScreen
 import com.example.androiddevchallenge.ui.theme.MyTheme
 
 class MainActivity : AppCompatActivity() {
@@ -63,26 +56,5 @@ fun LightPreview() {
 fun DarkPreview() {
     MyTheme(darkTheme = true) {
         MyApp()
-    }
-}
-
-@Composable
-fun NavGraph(startDestination: String = "cats") {
-    val navController = rememberNavController()
-
-    NavHost(
-        navController = navController,
-        startDestination = startDestination
-    ) {
-        composable("cats") { CatListScreen(navController) }
-        composable(
-            "cat/{catId}",
-            arguments = listOf(navArgument("catId") { type = NavType.IntType })
-        ) { backStackEntry ->
-            CatDetailScreen(
-                navController,
-                requireNotNull(backStackEntry.arguments).getInt("catId")
-            )
-        }
     }
 }
